requirements.txt
Flask==3.0.0
Flask-WTF==1.2.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
python-dotenv==1.0.0
email_validator==2.1.0.post1
psycopg2-binary==2.9.9
gunicorn==21.2.0
